configure auth in server

+++ server/src/routes/memories.ts
@@ -4,11 +4,14 @@ import { z } from 'zod';
 import { prisma } from '../lib/prisma';
 
 export async function memoriesRoutes(app: FastifyInstance) {
-	app.get('/memories', async () => {
+	app.addHook('preHandler', async (request) => {
+		await request.jwtVerify();
+	});
+
+	app.get('/memories', async (request) => {
 		const memories = await prisma.memory.findMany({
-			orderBy: {
-				createdAt: 'asc'
-			}
+			where: { userId: request.user.sub },
+			orderBy: { createdAt: 'asc' }
 		});
 
 		return memories.map(memory => ({
@@ -18,7 +21,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		}));
 	});
 
-	app.get('/memories/:id', async (request) => {
+	app.get('/memories/:id', async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid()
 		});
@@ -28,6 +31,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		const memory = await prisma.memory.findUniqueOrThrow({
 			where: { id }
 		});
+
+		if (!memory.isPublic && memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
 
 		return memory;
 	});
@@ -46,14 +53,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
 				content,
 				coverUrl,
 				isPublic,
-				userId: 'c487f79d-5562-4c59-84b3-c1d73a5a1229'
+				userId: request.user.sub
 			}
 		});
 
 		return memory;
 	});
 
-	app.put('/memories/:id', async (request) => {
+	app.put('/memories/:id', async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid()
 		});
@@ -68,7 +75,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
 		const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-		const memory = await prisma.memory.update({
+		let memory = await prisma.memory.findUniqueOrThrow({
+			where: { id }
+		});
+
+		if (memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
+
+		memory = await prisma.memory.update({
 			where: { id },
 			data: {
 				content,
@@ -80,12 +95,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		return memory;
 	});
 
-	app.delete('/memories/:id', async (request) => {
+	app.delete('/memories/:id', async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid()
 		});
 
 		const { id } = paramsSchema.parse(request.params);
+
+		const memory = await prisma.memory.findUniqueOrThrow({
+			where: { id }
+		});
+
+		if (memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
 
 		await prisma.memory.delete({
 			where: { id }
